seperate device button dialogs into their own files for easier debugging
